add rating and add search form

+++ db/migrate/20160512034429_add_rating_to_pitches.rb
@@ -1,5 +1,5 @@
 class AddRatingToPitches < ActiveRecord::Migration
   def change
-    add_column :pitches, :rating, :integer
+    add_column :pitches, :rating, :float
   end
 end
